A LOT OF THINGS CHANGED
i made it so theres now 20 different ground types with the obticles, powerups, and coins baked in. the golem also got a bit of an update. it resets when it gets too far or falls off the map, it also speeds up a bit when too far from the player. Some other bugs were fixed of course.

The terrain generation was fixed so now those terrain blocks arnt stacked on top of each other.